+++ CHANGELOG.md
@@ -7,6 +7,13 @@ Changes to this project will be documented in this file.
 ### Removed
 ### Fixed
 
+## [0.0.2]
+### Added
+- Support for "optional" project parts can be marked as "excluded" in project builds
+### Fixed
+- Bill of materials link in project version page
+- Fixes to test_project
+
 
 ## [0.0.1]
 - Initial release with: